Some initial refactoring, restructuring and commenting out (for now)
