+++ SassAndCoffee.Core/ICompilerHost.cs
@@ -1,0 +1,7 @@
+namespace SassAndCoffee.Core
+{
+    public interface ICompilerHost
+    {
+        ISimpleFileCompiler MapPathToCompiler(string physicalPath);
+    }
+}
