+++ SassAndCoffee.Core/ISimpleFileCompiler.cs
@@ -1,0 +1,13 @@
+namespace SassAndCoffee.Core
+{
+    public interface ISimpleFileCompiler
+    {
+        string[] InputFileExtensions { get; }
+        string OutputFileExtension { get; }
+        string OutputMimeType { get; }
+
+        void Init();
+        string ProcessFileContent(string inputFileContent);
+        string GetFileChangeToken(string inputFileContent);
+    }
+}
