+++ WebTest/Web.config
@@ -68,14 +68,14 @@
       </namespaces>
     </pages>
     <httpModules>
-      <add name="CompilableFileModule" type="SassAndCoffee.CompilableFileModule" />
+      <add name="CompilableFileModule" type="SassAndCoffee.AspNet.CompilableFileModule" />
     </httpModules>
   </system.web>
 
   <system.webServer>
     <validation validateIntegratedModeConfiguration="false" />
     <modules runAllManagedModulesForAllRequests="true">
-      <add name="CompilableFileModule" type="SassAndCoffee.CompilableFileModule" />
+      <add name="CompilableFileModule" type="SassAndCoffee.AspNet.CompilableFileModule" />
     </modules>
   </system.webServer>
